Adding documentation and moving coordinate to its own file
